+++ packages/localnet/src/tokens/createSolanaToken.ts
@@ -1,0 +1,129 @@
+import { BN } from "@coral-xyz/anchor";
+import {
+  createMint,
+  getOrCreateAssociatedTokenAccount,
+  mintTo,
+} from "@solana/spl-token";
+import { LAMPORTS_PER_SOL, PublicKey, SystemProgram } from "@solana/web3.js";
+
+import { ed25519KeyPairTSS as tssKeypair } from "../solanaSetup";
+
+export const createSolanaToken = async (env: any) => {
+  const mint = await createMint(
+    env.gatewayProgram.provider.connection,
+    tssKeypair,
+    tssKeypair.publicKey,
+    null,
+    9
+  );
+
+  const GATEWAY_PROGRAM_ID = env.gatewayProgram.programId;
+
+  const [gatewayPDA] = await PublicKey.findProgramAddress(
+    [Buffer.from("meta")],
+    GATEWAY_PROGRAM_ID
+  );
+
+  const [gatewayTokenAccount, tssTokenAccount, userTokenAccount] =
+    await Promise.all([
+      getOrCreateAssociatedTokenAccount(
+        env.gatewayProgram.provider.connection,
+        tssKeypair,
+        mint,
+        gatewayPDA,
+        true // allowOwnerOffCurve = true, because gatewayPDA is a program-derived address
+      ),
+      getOrCreateAssociatedTokenAccount(
+        env.gatewayProgram.provider.connection,
+        tssKeypair,
+        mint,
+        tssKeypair.publicKey
+      ),
+      getOrCreateAssociatedTokenAccount(
+        env.gatewayProgram.provider.connection,
+        env.defaultSolanaUser,
+        mint,
+        env.defaultSolanaUser.publicKey
+      ),
+    ]);
+
+  await Promise.all([
+    mintTo(
+      env.gatewayProgram.provider.connection,
+      tssKeypair,
+      mint,
+      tssTokenAccount.address,
+      tssKeypair.publicKey,
+      100 * LAMPORTS_PER_SOL
+    ),
+    mintTo(
+      env.gatewayProgram.provider.connection,
+      tssKeypair,
+      mint,
+      userTokenAccount.address,
+      tssKeypair.publicKey,
+      100 * LAMPORTS_PER_SOL
+    ),
+
+    mintTo(
+      env.gatewayProgram.provider.connection,
+      tssKeypair,
+      mint,
+      gatewayTokenAccount.address,
+      tssKeypair.publicKey,
+      100 * LAMPORTS_PER_SOL
+    ),
+    whitelistSPLToken(env.gatewayProgram, mint, env.defaultSolanaUser),
+  ]);
+
+  return [
+    mint.toBase58(),
+    gatewayTokenAccount.address.toBase58(),
+    userTokenAccount.address.toBase58(),
+  ];
+};
+
+const whitelistSPLToken = async (
+  gatewayProgram: any,
+  mintPublicKey: PublicKey,
+  authorityKeypair: any
+) => {
+  const [gatewayPDA] = await PublicKey.findProgramAddress(
+    [Buffer.from("meta", "utf-8")],
+    gatewayProgram.programId
+  );
+
+  const pdaAccountData = await gatewayProgram.account.pda.fetch(gatewayPDA);
+  console.log("🚀 Gateway PDA Authority:", pdaAccountData.authority.toBase58());
+
+  if (!pdaAccountData.authority.equals(authorityKeypair.publicKey)) {
+    console.error(
+      "❌ Error: The provided signer is NOT the authority of the Gateway PDA."
+    );
+    process.exit(1);
+  }
+
+  const [whitelistEntryPDA] = await PublicKey.findProgramAddress(
+    [Buffer.from("whitelist"), mintPublicKey.toBuffer()],
+    gatewayProgram.programId
+  );
+
+  console.log(
+    "Whitelisting SPL Token. Whitelist Entry PDA:",
+    whitelistEntryPDA.toBase58()
+  );
+
+  await gatewayProgram.methods
+    .whitelistSplMint(new Uint8Array(64).fill(0), 0, [], new BN(0))
+    .accounts({
+      authority: authorityKeypair.publicKey,
+      pda: gatewayPDA,
+      systemProgram: SystemProgram.programId,
+      whitelistCandidate: mintPublicKey,
+      whitelistEntry: whitelistEntryPDA,
+    })
+    .signers([authorityKeypair])
+    .rpc();
+
+  console.log(`✅ Whitelisted SPL Token: ${mintPublicKey.toBase58()}`);
+};
